menu: handle muliple input files
Update the argument parse logic to handle any number of input files.
The idea is that the surrounding logic should be able to pass any number
of input JSON files, and the menu will do the right thing.

There are 4 files of interest, Server and Desktop variants of releases
and cdimage streams.  There is enough info in those streams so that we
can process them correctly, so passing the files in without further
context will work out.