menu: create criteria table
Given a simple stream JSON, some artifacts are interesting and some are
not.  By looking at the content_id in that JSON, we can later filter on
the exact product of interest without hardcoding on a given product id,
and be flexible about what to show based on that input files have been
supplied.